Reduce warning.

+++ src/Particle/HDFWalkerInput_0_4.cpp
@@ -163,7 +163,7 @@ bool HDFWalkerInput_0_4::read_hdf5(std::string h5name)
 
   typedef std::vector<QMCTraits::RealType> Buffer_t;
   Buffer_t posin;
-  std::array<int, 3> dims{nw_in, targetW.getTotalNum(), OHMMS_DIM};
+  std::array<int, 3> dims{nw_in, static_cast<int>(targetW.getTotalNum()), OHMMS_DIM};
   posin.resize(dims[0] * dims[1] * dims[2]);
   hin.readSlabReshaped(posin, dims, hdf::walkers);
 
@@ -258,7 +258,7 @@ bool HDFWalkerInput_0_4::read_hdf5_scatter(std::string h5name)
   for (int i = 0; i < counts.size(); ++i)
     counts[i] = woffsets[i + 1] - woffsets[i];
 
-  std::array<int, 3> dims{nw_in, targetW.getTotalNum(), OHMMS_DIM};
+  std::array<int, 3> dims{nw_in, static_cast<int>(targetW.getTotalNum()), OHMMS_DIM};
   Buffer_t posin(nw_in * nitems), posout(counts[myComm->rank()]);
 
   if (myComm->rank() == 0)
@@ -342,7 +342,7 @@ bool HDFWalkerInput_0_4::read_phdf5(std::string h5name)
 
   typedef std::vector<QMCTraits::RealType> Buffer_t;
   Buffer_t posin;
-  std::array<int, 3> dims{nw_in, targetW.getTotalNum(), OHMMS_DIM};
+  std::array<int, 3> dims{nw_in, static_cast<int>(targetW.getTotalNum()), OHMMS_DIM};
 
   if (woffsets.size() != myComm->size() + 1)
   {
@@ -352,7 +352,7 @@ bool HDFWalkerInput_0_4::read_phdf5(std::string h5name)
 
   int nw_loc = woffsets[myComm->rank() + 1] - woffsets[myComm->rank()];
 
-  std::array<int, 3> counts{nw_loc, targetW.getTotalNum(), OHMMS_DIM};
+  std::array<int, 3> counts{nw_loc, static_cast<int>(targetW.getTotalNum()), OHMMS_DIM};
   std::array<int, 3> offsets{woffsets[myComm->rank()], 0, 0};
   posin.resize(nw_loc * dims[1] * dims[2]);
